DatePicker renamed to DatePickerInput
